Add files via upload
Problem Statement

What is the Problem?
It is critical to learn about the challenges with the current remittance system to comprehend the use cases of blockchain in the remittance business. Because most remittance sectors rely on third-party middlemen and financial institutions, the existing system is inefficient. Second, remittances paid through standard banking channels require the recipient to have a bank account to receive the payment. But what if the recipients of the funds do not have a bank account? Furthermore, such traditional services are costly and time-consuming, as funds transmission might take days or even weeks.
Solution of the Problem.
We are going to create a web-based smart System which is based on BLOCKCHAIN ENABLES CROSS-BORDER REMITTANCE SYSTEM which not only shows the details of the transaction but it will also manage the fess and time duration of transaction. You should only register in our system to enjoy fastest transaction system. We also have added the functionality of seeing the transaction details by which you are able to see to whom you have done transaction with.


Objectives

Blockchain technology, unlike traditional remittance techniques, completely transforms the way foreign payments are made. This is because blockchain provides a fast, transparent, and secure way to make international payments. It uses a distributed network of computers to conduct global financial transactions. It can streamline the entire transaction process by eliminating redundant intermediaries and providing worldwide payment options practically instantaneously. By the motivation, as mentioned earlier, this thesis is going to design a novel blockchain-enabled cross border remittance system in distributed computing nodes. The study has the following objectives.
•	Design the novel blockchain-enabled cross border remittance system for digital currency transferring.
•	Design the novel distributed ledger system for banking based on blockchain technology. 
•	The considered problem is categorized as the combinatorial optimation problem, where cost, time and security are to be reduced as the conflicting objectives with the set of convex set constraints.
•	The study design a novel use-cases flow diagram based on fully dressed model, where all processes to be discussed step by step.
Significance of Problem

Since the emphasis of this project is to make the system decentralized that make it secure, time saving and cost efficient. It will be exceptionally significant for investigators in transforming their systems into a convenient environment. It will increase effectiveness by making it easier to identify data sources correctly early in the investigation process, as a result of this project a highly distributes web-application will be developed

Limitation of Problem

Our Project milestone is to make a smart transaction system that has advanced features including Cost, security, scalability, and efficiency, Lack of standardization, Improved record transparency, Regulatory compliances, Strengthened Chain of processing with transaction etc. 
Proposed Project
Details of Processing while developing this project
Requirements:
There are certain requirements for this project and those requirements are mostly for financial industries and those who do transaction. It means that project first priority will be the interest of the financial institutes. The second requirement is process of transaction because transaction matters for those who mostly transfer fund and have not bank account. Third requirement is interconnecting of banks through nodes directly to each other. In many cases user who do transaction can’t afford the charges he faced for doing transaction from one state to another state so our proposed system will make all of these factor better.
Design:
Keeping an eye on the requirements and after analyzing that requirements that are going to be designed our project accordingly. In this phase remittance system for cross-border will be developed for the sake of the financial industry and that recommender system will base on blockchain cross-border remittance system. In blockchain system a famous technique for decentralization and hyperledger will be performed.
Development:
In this phase we will write code for our project. This phase totally depends on the coding of the system. How UI will look like, which features should be added for user registration, with method to follow for connectivity of our web-app with blockchain and so on.
Testing:
In this phase testing of entire system will be focused. We test each and every faults and failures of our system in different. Firstly, we are going to perform unit testing for each module developed that is it working fine or not. Later on after complete development of our entire system we are going to perform system testing also is complete system is working fine or not.  If our entire system is working properly and fulfilling all the requirements of our system, then it moves towards the next phase which is deployment.
Deployment:
In this phase we will make surety that the code and the technology which we have applied on our system should be appropriate and up to the date because in this phase we are going to launch our system in the market sector, so we have to be very much sure of our code and technology.
Maintenance:
In this phase, the system should be updated with the passage of time. This phase also includes the satisfaction of the users.

Implementation

Flask is a web framework. This means flask provides you with tools, libraries and technologies that allow you to build a web application. This web application can be some web pages, a blog, a wiki or go as big as a web-based calendar application or a commercial website.

Blockchain is a system of recording information in a way that makes it difficult or impossible to change, hack, or cheat the system. A blockchain is essentially a digital ledger of transactions that is duplicated and distributed across the entire network of computer systems on the blockchain.

MySQL is a popular, free-to-use, and open-source relational database management system (RDBMS) developed by Oracle. As with other relational systems, MySQL stores data using tables and rows, enforces referential integrity and uses structured query language (SQL) for data access.